Not to free timer object
