+++ src/api_manager/config_manager.cc
@@ -44,6 +44,9 @@ ConfigManager::ConfigManager(
                                         .fetch_throttle_window_s();
     }
   }
+  static std::random_device random_device;
+  random_generator_.seed(random_device());
+
   // throttle in milliseconds
   random_dist_.reset(new std::uniform_int_distribution<int>(
       0, fetch_throttle_window_in_s_ * 1000));
@@ -64,8 +67,8 @@ void ConfigManager::SetLatestRolloutId(
     return;
   }
 
-  // Timer is pending, or too close to last fetch time
-  if (fetch_timer_ || now < next_window_time_) {
+  // Last timer is not fired yet,  or is too close to last fetch time
+  if ((fetch_timer_ && !fetch_timer_->IsStopped()) || now < next_window_time_) {
     return;
   }
 
@@ -79,9 +82,7 @@ void ConfigManager::SetLatestRolloutId(
 
 void ConfigManager::OnRolloutsRefreshTimer() {
   global_context_->env()->LogInfo("Fetch timer starts");
-
   fetch_timer_->Stop();
-  fetch_timer_ = nullptr;
   next_window_time_ = std::chrono::system_clock::now() +
                       std::chrono::seconds(fetch_throttle_window_in_s_);
