+++ src/api_manager/context/BUILD
@@ -44,6 +44,7 @@ cc_library(
         "//external:service_config",
         "//external:servicecontrol",
         "//external:servicecontrol_client",
+        "//src/api_manager:compute_platform",
         "//src/api_manager:http_template",
         "//src/api_manager:impl_headers",
         "//src/api_manager:path_matcher",
@@ -52,7 +53,6 @@ cc_library(
         "//src/api_manager/auth:authz",
         "//src/api_manager/auth:service_account_token",
         "//src/api_manager/cloud_trace",
-        "//src/api_manager:compute_platform",
         "//src/api_manager/service_control",
         "//src/api_manager/utils",
     ],
@@ -67,6 +67,7 @@ cc_test(
     deps = [
         "//external:googletest_main",
         "//src/api_manager",
+        "//src/api_manager:mock_api_manager_environment",
     ],
 )
